requirements.txt
# Core training stack
torch>=2.1.0
transformers==4.51.3
datasets>=2.18.0
accelerate>=0.28.0
trl>=0.8.6
peft>=0.10.0
bitsandbytes>=0.43.0
sentencepiece>=0.1.99
protobuf>=4.25.0
evaluate>=0.4.0

# Quantization & inference
llama-cpp-python>=0.2.55
gguf>=0.6.0

# Demo
gradio>=4.26.0

# Utilities
matplotlib>=3.8.0
seaborn>=0.13.0
numpy>=1.26.0
